+++ src/test/java/com/github/alexisjehan/javanilla/lang/array/IntArraysTest.java
@@ -86,16 +86,33 @@ final class IntArraysTest {
 	@Test
 	void testAdd() {
 		assertThat(IntArrays.add(IntArrays.EMPTY, 0)).containsExactly(0);
-		assertThat(IntArrays.addTemporary(IntArrays.of(1, 2, 3), 0, 0)).containsExactly(0, 1, 2, 3);
-		assertThat(IntArrays.addTemporary(IntArrays.of(1, 2, 3), 0, 1)).containsExactly(1, 0, 2, 3);
-		assertThat(IntArrays.addTemporary(IntArrays.of(1, 2, 3), 0, 2)).containsExactly(1, 2, 0, 3);
-		assertThat(IntArrays.addTemporary(IntArrays.of(1, 2, 3), 0, 3)).containsExactly(1, 2, 3, 0);
+		assertThat(IntArrays.add(IntArrays.of(1, 2, 3), 0, 0)).containsExactly(0, 1, 2, 3);
+		assertThat(IntArrays.add(IntArrays.of(1, 2, 3), 0, 1)).containsExactly(1, 0, 2, 3);
+		assertThat(IntArrays.add(IntArrays.of(1, 2, 3), 0, 2)).containsExactly(1, 2, 0, 3);
+		assertThat(IntArrays.add(IntArrays.of(1, 2, 3), 0, 3)).containsExactly(1, 2, 3, 0);
 		assertThat(IntArrays.add(IntArrays.of(1, 2, 3), 0)).containsExactly(1, 2, 3, 0);
 	}
 
 	@Test
 	void testAddInvalid() {
 		assertThatNullPointerException().isThrownBy(() -> IntArrays.add(null, 0));
+		assertThatNullPointerException().isThrownBy(() -> IntArrays.add(null, 0, 0));
+		assertThatIllegalArgumentException().isThrownBy(() -> IntArrays.add(IntArrays.of(VALUES), 0, -1));
+		assertThatIllegalArgumentException().isThrownBy(() -> IntArrays.add(IntArrays.of(VALUES), 0, 3));
+	}
+
+	@Test
+	@SuppressWarnings("deprecation")
+	void testAddTemporary() {
+		assertThat(IntArrays.addTemporary(IntArrays.of(1, 2, 3), 0, 0)).containsExactly(0, 1, 2, 3);
+		assertThat(IntArrays.addTemporary(IntArrays.of(1, 2, 3), 0, 1)).containsExactly(1, 0, 2, 3);
+		assertThat(IntArrays.addTemporary(IntArrays.of(1, 2, 3), 0, 2)).containsExactly(1, 2, 0, 3);
+		assertThat(IntArrays.addTemporary(IntArrays.of(1, 2, 3), 0, 3)).containsExactly(1, 2, 3, 0);
+	}
+
+	@Test
+	@SuppressWarnings("deprecation")
+	void testAddTemporaryInvalid() {
 		assertThatNullPointerException().isThrownBy(() -> IntArrays.addTemporary(null, 0, 0));
 		assertThatIllegalArgumentException().isThrownBy(() -> IntArrays.addTemporary(IntArrays.of(VALUES), 0, -1));
 		assertThatIllegalArgumentException().isThrownBy(() -> IntArrays.addTemporary(IntArrays.of(VALUES), 0, 3));